wifi: fix wifi close issue
PD# 174173

root cause: interrupt not enabled when ctrl cmd comming in some conner case
solution: enable the interrupt when ctrl command is comming

+++ bcmdhd.1.579.77.41.1.cn/dbus.c
@@ -1442,7 +1442,7 @@ dbus_attach(osl_t *osh, int rxsize, int nrxq, int ntxq, dhd_pub_t *pub,
 
 	dhd_bus = MALLOC(osh, sizeof(dhd_bus_t));
 	if (dhd_bus == NULL) {
-		DBUSERR(("%s: malloc failed %d\n", __FUNCTION__, (int)sizeof(dhd_bus_t)));
+		DBUSERR(("%s: malloc failed %zu\n", __FUNCTION__, sizeof(dhd_bus_t)));
 		return NULL;
 	}
